# openapi.yaml
openapi: 3.0.3
info:
  title: EvolutionHub API
  version: "1.0.0"
  description: Export der Routen aus src/pages als OpenAPI 3.0 Spec
servers:
  - url: /
    description: Default server (relative)
paths:
...
  /api/admin/backup/jobs:
    get:
      summary: List backup jobs
      description: Returns recent backup jobs ordered by startedAt desc. Requires admin.
      x-source: src/pages/api/admin/backup/jobs/index.ts
      operationId: getAdminBackupJobs
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "401":
          description: Unauthorized
  /api/admin/backup/jobs/{id}:
    get:
      summary: Get backup job
      description: Returns details for a specific backup job. Requires admin.
      x-source: src/pages/api/admin/backup/jobs/[id]/index.ts
      operationId: getAdminBackupJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "404":
          description: Not found
  /api/admin/backup/jobs/{id}/progress:
    get:
      summary: Get backup progress
      description: Returns progress for a running or completed backup job. Requires admin.
      x-source: src/pages/api/admin/backup/jobs/[id]/progress.ts
      operationId: getAdminBackupJobProgress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Progress object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "404":
          description: Not found
  /api/admin/backup/maintenance/jobs:
    get:
      summary: List maintenance jobs
      description: Returns recent maintenance jobs ordered by startedAt desc. Requires admin.
      x-source: src/pages/api/admin/backup/maintenance/jobs/index.ts
      operationId: getAdminMaintenanceJobs
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
  /api/admin/backup/maintenance/jobs/{id}:
    get:
      summary: Get maintenance job
      description: Returns details for a specific maintenance job. Requires admin.
      x-source: src/pages/api/admin/backup/maintenance/jobs/[id].ts
      operationId: getAdminMaintenanceJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "404":
          description: Not found
...
  /api/admin/backup/create:
    post:
      summary: Create backup job
      description: Creates a new backup job. Requires admin.
      x-source: src/pages/api/admin/backup/create.ts
      operationId: postAdminBackupCreate
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          description: Double‑Submit token; must match cookie 'csrf_token'
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                tables:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Backup job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "403":
          description: Forbidden (CSRF/Origin or insufficient permissions)
  /api/admin/backup/schedule:
    post:
      summary: Schedule automated backup
      description: Schedules an automated backup based on a cron expression. Requires admin.
      x-source: src/pages/api/admin/backup/schedule.ts
      operationId: postAdminBackupSchedule
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          description: Double‑Submit token; must match cookie 'csrf_token'
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                cronExpression:
                  type: string
      responses:
        "200":
          description: Scheduled
  /api/admin/backup/maintenance/perform:
    post:
      summary: Perform system maintenance
      description: Starts a maintenance job. Requires admin.
      x-source: src/pages/api/admin/backup/maintenance/perform.ts
      operationId: postAdminMaintenancePerform
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          description: Double‑Submit token; must match cookie 'csrf_token'
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [cleanup, optimization, migration, repair]
                description:
                  type: string
      responses:
        "200":
          description: Maintenance started
  /api/admin/backup/cleanup:
    post:
      summary: Cleanup old backups
      description: Deletes old completed backups based on retention policy. Requires admin.
      x-source: src/pages/api/admin/backup/cleanup.ts
      operationId: postAdminBackupCleanup
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          description: Double‑Submit token; must match cookie 'csrf_token'
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                retentionDays:
                  type: integer
      responses:
        "200":
          description: Cleanup result
  /api/admin/backup/verify/{id}:
    post:
      summary: Verify backup integrity
      description: Verifies integrity of a completed backup by ID. Requires admin.
      x-source: src/pages/api/admin/backup/verify/[id].ts
      operationId: postAdminBackupVerify
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: X-CSRF-Token
          in: header
          required: true
          description: Double‑Submit token; must match cookie 'csrf_token'
          schema:
            type: string
      responses:
        "200":
          description: Verification result
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
    schemas:
...
      GenericResponse:
        type: object
        properties:
          success:
            type: boolean
...
      ApiError:
        type: object
        properties:
          type:
            type: string
          message:
            type: string
          details:
            type: object
            additionalProperties: true
      ApiErrorResponse:
        type: object
        properties:
          success:
            type: boolean
            enum: [false]
          error:
            $ref: '#/components/schemas/ApiError'
        required:
          - success
          - error
...